Moving stan/io/mcmc_writer.hpp to stan/services/sample/mcmc_writer.hpp
