+++ src/stan/services/mcmc/sample.hpp
@@ -2,7 +2,7 @@
 #define STAN_SERVICES_MCMC_SAMPLE_HPP
 
 #include <stan/mcmc/base_mcmc.hpp>
-#include <stan/io/mcmc_writer.hpp>
+#include <stan/services/sample/mcmc_writer.hpp>
 #include <stan/services/mcmc/run_markov_chain.hpp>
 #include <string>
 
@@ -19,9 +19,9 @@ namespace stan {
                   int num_thin,
                   int refresh,
                   bool save,
-                  stan::io::mcmc_writer<
-                    Model, SampleRecorder, DiagnosticRecorder, MessageRecorder>&
-                    writer,
+                  stan::services::sample::mcmc_writer<
+                  Model, SampleRecorder, DiagnosticRecorder, MessageRecorder>&
+                  writer,
                   stan::mcmc::sample& init_s,
                   Model& model,
                   RNG& base_rng,
@@ -31,13 +31,13 @@ namespace stan {
                   StartTransitionCallback& callback) {
         run_markov_chain<Model, RNG, StartTransitionCallback,
                          SampleRecorder, DiagnosticRecorder, MessageRecorder>(
-           sampler, num_samples, num_warmup,
-           num_warmup + num_samples, num_thin,
-           refresh, save, false,
-           writer,
-           init_s, model, base_rng,
-           prefix, suffix, o,
-           callback);
+                                                                              sampler, num_samples, num_warmup,
+                                                                              num_warmup + num_samples, num_thin,
+                                                                              refresh, save, false,
+                                                                              writer,
+                                                                              init_s, model, base_rng,
+                                                                              prefix, suffix, o,
+                                                                              callback);
       }
 
     }
